backend/requirements.txt
absl-py==2.3.1
aiosqlite==0.22.1
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
bcrypt==4.0.1
billiard==4.2.4
blinker==1.9.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
cachetools==6.2.3
celery==5.6.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1762976168352/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761202952988/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm==0.2.3
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1764077554327/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1764081326783/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databases==0.9.0
debugpy==1.8.19
decorator==5.2.1
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.129.0
filelock==3.24.3
filetype==1.2.0
Flask==3.1.3
flatbuffers==25.9.23
fonttools==4.61.1
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1763082885869/work
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
h5py==3.15.1
haystack-ai==2.24.1
haystack-experimental==0.19.0
hf-xet==1.3.2
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1756754188614/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kagglehub==0.3.13
keras==3.12.0
kiwisolver==1.4.9
kombu==5.6.2
lazy_imports==1.2.0
libclang==18.1.1
libmambapy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_libmambapy_1764158592/work/libmambapy
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1761299914766/work
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1762504198966/work
namex==0.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.0
ollama==0.6.1
ollama-haystack==6.0.0
openai==2.21.0
opt_einsum==3.4.0
optree==0.18.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
pluggy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pluggy_1764896838/work
posthog==7.8.6
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1757744682161/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
pypdf==6.7.4
PyPDF2==3.0.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1761160761767/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1760564263846/work
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.47
stack-data==0.6.3
starlette==0.52.1
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
transformers==5.2.0
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1753886790/work
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.6
wheel==0.45.1
word2number==1.1
wrapt==2.0.1
zstandard==0.25.0
